HomePage and DashboardPage turned into client components
